bracketiq/requirements.txt
# BracketIQ backend dependencies
# KenPom scraper (install from PyPI or GitHub)
# https://github.com/j-andrews7/kenpompy
kenpompy>=0.3.0

# API & server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data & validation
pandas>=2.0.0
pyarrow>=14.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP client for The Odds API
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
scipy>=1.11.0

# Neon PostgreSQL (push_to_neon, pull_from_neon)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
